passport-ocr-app/backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
pytesseract==0.3.10
passporteye==2.1.0
opencv-python==4.8.1.78
Pillow==10.1.0
numpy==1.24.3

